requirements.txt
# Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
mysqlclient>=2.2.0
pymysql>=1.1.0
cryptography>=41.0.5  # Required for PyMySQL
alembic>=1.12.1
httpx>=0.24.0